refactor(okam-build): move syntax plugin to syntax dir
