+++ packages/okam-build/lib/processor/template/plugins/syntax/ant-syntax-plugin.js
@@ -5,7 +5,7 @@
 
 'use strict';
 
-const {createSyntaxPlugin} = require('./helper');
-const transformers = require('../transform/ant');
+const {createSyntaxPlugin} = require('../helper');
+const transformers = require('../../transform/ant');
 
 module.exports = createSyntaxPlugin(transformers);
