+++ packages/okam-build/lib/processor/template/plugins/syntax/quick-syntax-plugin.js
@@ -5,8 +5,8 @@
 
 'use strict';
 
-const {createSyntaxPlugin} = require('./helper');
-const transformers = require('../transform/quick');
+const {createSyntaxPlugin} = require('../helper');
+const transformers = require('../../transform/quick');
 
 module.exports = createSyntaxPlugin(transformers);
