+++ packages/okam-build/lib/processor/template/plugins/syntax/wx2swan-syntax-plugin.js
@@ -5,8 +5,8 @@
 
 'use strict';
 
-const {createSyntaxPlugin} = require('./helper');
-const transformers = require('../transform/wx2swan');
+const {createSyntaxPlugin} = require('../helper');
+const transformers = require('../../transform/wx2swan');
 
 module.exports = createSyntaxPlugin(transformers);
